Cleanup usage of Hashable protocol.
