+++ Sources/Client/Model/Message/Attachment/Attachment.swift
@@ -176,8 +176,7 @@ public struct Attachment: Codable {
     }
 }
 
-extension Attachment: Hashable {
-    
+extension Attachment: Equatable {
     public static func == (lhs: Attachment, rhs: Attachment) -> Bool {
         lhs.title == rhs.title
             && lhs.author == rhs.author
@@ -186,16 +185,6 @@ extension Attachment: Hashable {
             && lhs.url == rhs.url
             && lhs.imageURL == rhs.imageURL
             && lhs.file == rhs.file
-    }
-    
-    public func hash(into hasher: inout Hasher) {
-        hasher.combine(title)
-        hasher.combine(author)
-        hasher.combine(text)
-        hasher.combine(type)
-        hasher.combine(url)
-        hasher.combine(imageURL)
-        hasher.combine(file)
     }
 }
