+++ Sources/Client/Model/Message/Attachment/AttachmentType.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 /// An attachment type.
-public enum AttachmentType: RawRepresentable, Codable, Hashable, ExpressibleByStringLiteral {
+public enum AttachmentType: RawRepresentable, Codable, ExpressibleByStringLiteral {
     /// An attachment type.
     case unknown
     case custom(String)
