+++ Sources/Client/Model/UnreadCount.swift
@@ -7,14 +7,14 @@
 //
 
 /// An unread counts of the current user.
-public struct UnreadCount: Decodable, Hashable {
+public struct UnreadCount: Decodable, Equatable {
     public static let noUnread = UnreadCount(channels: 0, messages: 0)
     public internal(set) var channels: Int
     public internal(set) var messages: Int
 }
 
 /// An unread counts for a channel.
-public struct ChannelUnreadCount: Decodable, Hashable {
+public struct ChannelUnreadCount: Decodable, Equatable {
     public static let noUnread = ChannelUnreadCount(messages: 0, mentionedMessages: 0)
     public internal(set) var messages: Int
     public internal(set) var mentionedMessages: Int
