+++ Sources/UI/Extensions/UIKit/UIImage/UIImage+Render.swift
@@ -180,10 +180,3 @@ extension UIImage {
         return path
     }
 }
-
-extension UIRectCorner: Hashable {
-    public static let leftSide: UIRectCorner = [.topLeft, .bottomLeft]
-    public static let rightSide: UIRectCorner = [.topRight, .bottomRight]
-    public static let pointedLeftBottom: UIRectCorner = [.topLeft, .topRight, .bottomRight]
-    public static let pointedRightBottom: UIRectCorner = [.topLeft, .topRight, .bottomLeft]
-}
