+++ Sources/UI/Style/ChannelViewStyle.swift
@@ -130,53 +130,3 @@ public struct ChannelViewStyle {
             : height
     }
 }
-
-extension ChannelViewStyle: Hashable {
-    
-    public static func == (lhs: ChannelViewStyle, rhs: ChannelViewStyle) -> Bool {
-        lhs.backgroundColor == rhs.backgroundColor
-            && lhs.separatorStyle == rhs.separatorStyle
-            && lhs.avatarViewStyle == rhs.avatarViewStyle
-            && lhs.nameNumberOfLines == rhs.nameNumberOfLines
-            && lhs.nameFont == rhs.nameFont
-            && lhs.nameColor == rhs.nameColor
-            && lhs.nameUnreadFont == rhs.nameUnreadFont
-            && lhs.nameUnreadColor == rhs.nameUnreadColor
-            && lhs.messageNumberOfLines == rhs.messageNumberOfLines
-            && lhs.messageFont == rhs.messageFont
-            && lhs.messageColor == rhs.messageColor
-            && lhs.messageUnreadFont == rhs.messageUnreadFont
-            && lhs.messageUnreadColor == rhs.messageUnreadColor
-            && lhs.messageDeletedFont == rhs.messageDeletedFont
-            && lhs.messageDeletedColor == rhs.messageDeletedColor
-            && lhs.dateFont == rhs.dateFont
-            && lhs.dateColor == rhs.dateColor
-            && lhs.spacing == rhs.spacing
-            && lhs.edgeInsets == rhs.edgeInsets
-    }
-    
-    public func hash(into hasher: inout Hasher) {
-        hasher.combine(backgroundColor)
-        hasher.combine(separatorStyle)
-        hasher.combine(avatarViewStyle)
-        hasher.combine(nameNumberOfLines)
-        hasher.combine(nameFont)
-        hasher.combine(nameColor)
-        hasher.combine(nameUnreadFont)
-        hasher.combine(nameUnreadColor)
-        hasher.combine(messageNumberOfLines)
-        hasher.combine(messageFont)
-        hasher.combine(messageColor)
-        hasher.combine(messageUnreadFont)
-        hasher.combine(messageUnreadColor)
-        hasher.combine(messageDeletedFont)
-        hasher.combine(messageDeletedColor)
-        hasher.combine(dateFont)
-        hasher.combine(dateColor)
-        hasher.combine(spacing)
-        hasher.combine(edgeInsets.top)
-        hasher.combine(edgeInsets.left)
-        hasher.combine(edgeInsets.bottom)
-        hasher.combine(edgeInsets.right)
-    }
-}
