+++ Sources/UI/Style/MessageViewStyle.swift
@@ -250,133 +250,4 @@ public struct MessageViewStyle: Hashable {
                                                                    borderWidth: borderWidth,
                                                                    borderColor: borderColor)]
     }
-    
-    // MARK: - Hashable
-    
-    public static func == (lhs: MessageViewStyle, rhs: MessageViewStyle) -> Bool {
-        lhs.alignment == rhs.alignment
-            && lhs.avatarViewStyle == rhs.avatarViewStyle
-            && lhs.chatBackgroundColor == rhs.chatBackgroundColor
-            && lhs.font == rhs.font
-            && lhs.replyFont == rhs.replyFont
-            && lhs.nameFont == rhs.nameFont
-            && lhs.infoFont == rhs.infoFont
-            && lhs.emojiFont == rhs.emojiFont
-            && lhs.textColor == rhs.textColor
-            && lhs.replyColor == rhs.replyColor
-            && lhs.infoColor == rhs.infoColor
-            && lhs.backgroundColor == rhs.backgroundColor
-            && lhs.borderColor == rhs.borderColor
-            && lhs.borderWidth == rhs.borderWidth
-            && lhs.cornerRadius == rhs.cornerRadius
-            && lhs.pointedCornerRadius == rhs.pointedCornerRadius
-            && lhs.spacing == rhs.spacing
-            && lhs.edgeInsets == rhs.edgeInsets
-            && lhs.messageInsetSpacing == rhs.messageInsetSpacing
-            && lhs.reactionViewStyle == rhs.reactionViewStyle
-            && lhs.showTimeThreshold == rhs.showTimeThreshold
-            && lhs.additionalDateStyle == rhs.additionalDateStyle
-            && lhs.markdownEnabled == rhs.markdownEnabled
-    }
-    
-    public func hash(into hasher: inout Hasher) {
-        hasher.combine(alignment)
-        hasher.combine(avatarViewStyle)
-        hasher.combine(chatBackgroundColor)
-        hasher.combine(font)
-        hasher.combine(replyFont)
-        hasher.combine(nameFont)
-        hasher.combine(infoFont)
-        hasher.combine(emojiFont)
-        hasher.combine(textColor)
-        hasher.combine(replyColor)
-        hasher.combine(infoColor)
-        hasher.combine(backgroundColor)
-        hasher.combine(borderColor)
-        hasher.combine(borderWidth)
-        hasher.combine(cornerRadius)
-        hasher.combine(pointedCornerRadius)
-        hasher.combine(spacing)
-        hasher.combine(edgeInsets.top)
-        hasher.combine(edgeInsets.bottom)
-        hasher.combine(edgeInsets.left)
-        hasher.combine(edgeInsets.right)
-        hasher.combine(messageInsetSpacing)
-        hasher.combine(reactionViewStyle)
-        hasher.combine(showTimeThreshold)
-        hasher.combine(additionalDateStyle)
-        hasher.combine(markdownEnabled)
-    }
-}
-
-extension MessageViewStyle {
-    struct BackgroundImage {
-        private let images: [Int: UIImage]
-        private let defaultImage: UIImage?
-        
-        init(cornerRadius: CGFloat,
-             pointedCornerRadius: CGFloat,
-             corners: UIRectCorner,
-             color: UIColor,
-             backgroundColor: UIColor = .clear,
-             borderWidth: CGFloat = 0,
-             borderColor: UIColor = .black) {
-            if #available(iOS 13, *) {
-                let lightTrait = UITraitCollection(userInterfaceStyle: .light)
-                let darkTrait = UITraitCollection(userInterfaceStyle: .dark)
-                let lightColor = color.resolvedColor(with: lightTrait)
-                let darkColor = color.resolvedColor(with: darkTrait)
-                let lightBackgroundColor = backgroundColor.resolvedColor(with: lightTrait)
-                let darkBackgroundColor = backgroundColor.resolvedColor(with: darkTrait)
-                let lightBorderColor = borderColor.resolvedColor(with: lightTrait)
-                let darkBorderColor = borderColor.resolvedColor(with: darkTrait)
-                
-                if lightColor != darkColor || lightBackgroundColor != darkBackgroundColor || lightBorderColor != darkBorderColor {
-                    defaultImage = nil
-                    
-                    images = [UIUserInterfaceStyle.light.rawValue: .renderRounded(cornerRadius: cornerRadius,
-                                                                                  pointedCornerRadius: pointedCornerRadius,
-                                                                                  corners: corners,
-                                                                                  color: lightColor,
-                                                                                  backgroundColor: lightBackgroundColor,
-                                                                                  borderWidth: borderWidth,
-                                                                                  borderColor: lightBorderColor),
-                              
-                              UIUserInterfaceStyle.dark.rawValue: .renderRounded(cornerRadius: cornerRadius,
-                                                                                 pointedCornerRadius: pointedCornerRadius,
-                                                                                 corners: corners,
-                                                                                 color: darkColor,
-                                                                                 backgroundColor: darkBackgroundColor,
-                                                                                 borderWidth: borderWidth,
-                                                                                 borderColor: darkBorderColor)]
-                    
-                    return
-                }
-            }
-            
-            images = [:]
-            
-            defaultImage = .renderRounded(cornerRadius: cornerRadius,
-                                   pointedCornerRadius: pointedCornerRadius,
-                                   corners: corners,
-                                   color: color,
-                                   backgroundColor: backgroundColor,
-                                   borderWidth: borderWidth,
-                                   borderColor: borderColor)
-        }
-        
-        func image(for traitCollection: UITraitCollection) -> UIImage? {
-            guard #available(iOS 13, *), !images.isEmpty else {
-                return defaultImage
-            }
-            
-            switch traitCollection.userInterfaceStyle {
-            case .dark:
-                return images[UIUserInterfaceStyle.dark.rawValue]
-            default:
-                return images[UIUserInterfaceStyle.light.rawValue]
-            }
-            
-        }
-    }
 }
