+++ Sources/UI/Style/ReactionViewStyle.swift
@@ -9,7 +9,7 @@
 import UIKit
 
 /// A reaction view style.
-public struct ReactionViewStyle {
+public struct ReactionViewStyle: Hashable {
     
     /// An alignment of a reaction for incoming or outgoing messages.
     public let alignment: MessageViewStyle.Alignment
@@ -56,32 +56,5 @@ public struct ReactionViewStyle {
         self.tailCornerRadius = cornerRadius * 0.8
         self.tailMessageCornerRadius = tailMessageCornerRadius
         tailImage = .renderTailImage(smallRadius: tailCornerRadius, bigRadius: tailMessageCornerRadius, alignment: alignment)
-    }
-}
-
-extension ReactionViewStyle: Hashable {
-    
-    public static func == (lhs: ReactionViewStyle, rhs: ReactionViewStyle) -> Bool {
-        lhs.alignment == rhs.alignment
-            && lhs.font == rhs.font
-            && lhs.textColor == rhs.textColor
-            && lhs.backgroundColor == rhs.backgroundColor
-            && lhs.chatBackgroundColor == rhs.chatBackgroundColor
-            && lhs.cornerRadius == rhs.cornerRadius
-            && lhs.tailCornerRadius == rhs.tailCornerRadius
-            && lhs.tailMessageCornerRadius == rhs.tailMessageCornerRadius
-            && lhs.tailImage == rhs.tailImage
-    }
-    
-    public func hash(into hasher: inout Hasher) {
-        hasher.combine(alignment)
-        hasher.combine(font)
-        hasher.combine(textColor)
-        hasher.combine(backgroundColor)
-        hasher.combine(chatBackgroundColor)
-        hasher.combine(cornerRadius)
-        hasher.combine(tailCornerRadius)
-        hasher.combine(tailMessageCornerRadius)
-        hasher.combine(tailImage)
     }
 }
